@@ -1,10 +1,12 @@
 module github.com/coder/wush
 
-go 1.22.6
+go 1.23.1
 
-replace tailscale.com => github.com/coadler/tailscale v1.1.1-0.20240926000438-059d0c1039af
+toolchain go1.23.2
 
-// replace tailscale.com => /home/colin/Projects/coadler/tailscale
+// replace tailscale.com => github.com/coadler/tailscale v1.1.1-0.20240926000438-059d0c1039af
+
+replace tailscale.com => /home/colin/Projects/coadler/tailscale
 
 replace github.com/gliderlabs/ssh => github.com/coder/ssh v0.0.0-20231128192721-70855dedb788
 
