@@ -8,7 +8,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"math/rand/v2"
 	"net/netip"
 	"sync"
 	"sync/atomic"
@@ -16,6 +15,7 @@ import (
 	"time"
 
 	"github.com/coder/wush/cliui"
+	"github.com/pion/webrtc/v4"
 	"github.com/puzpuzpuz/xsync/v3"
 	"tailscale.com/derp"
 	"tailscale.com/derp/derphttp"
@@ -24,9 +24,15 @@ import (
 	"tailscale.com/net/portmapper"
 	"tailscale.com/tailcfg"
 	"tailscale.com/types/key"
+	"tailscale.com/types/logger"
 )
 
-func NewWasmOverlay(hlog Logf, dm *tailcfg.DERPMap, onNewPeer js.Value) *Wasm {
+func NewWasmOverlay(hlog Logf, dm *tailcfg.DERPMap,
+	onNewPeer js.Value,
+	onWebrtcOffer js.Value,
+	onWebrtcAnswer js.Value,
+	onWebrtcCandidate js.Value,
+) *Wasm {
 	return &Wasm{
 		HumanLogf: hlog,
 		DerpMap:   dm,
@@ -34,10 +40,13 @@ func NewWasmOverlay(hlog Logf, dm *tailcfg.DERPMap, onNewPeer js.Value) *Wasm {
 		PeerPriv:  key.NewNode(),
 		SelfIP:    randv6(),
 
-		peers:     xsync.NewMapOf[int32, chan<- *tailcfg.Node](),
-		onNewPeer: onNewPeer,
-		in:        make(chan *tailcfg.Node, 8),
-		out:       make(chan *tailcfg.Node, 8),
+		onNewPeer:         onNewPeer,
+		onWebrtcOffer:     onWebrtcOffer,
+		onWebrtcAnswer:    onWebrtcAnswer,
+		onWebrtcCandidate: onWebrtcCandidate,
+
+		in:  make(chan *tailcfg.Node, 8),
+		out: make(chan *overlayMessage, 8),
 	}
 }
 
@@ -61,16 +70,21 @@ type Wasm struct {
 	// DerpRegionID is the DERP region that can be used for proxied overlay
 	// communication.
 	DerpRegionID uint16
+	DerpLatency  time.Duration
 
 	// peers is a map of channels that notify peers of node updates.
-	peers     *xsync.MapOf[int32, chan<- *tailcfg.Node]
-	onNewPeer js.Value
+	activePeer atomic.Pointer[chan *overlayMessage]
+	onNewPeer  js.Value
+
+	onWebrtcOffer     js.Value
+	onWebrtcAnswer    js.Value
+	onWebrtcCandidate js.Value
 
 	lastNode atomic.Pointer[tailcfg.Node]
 	// in funnels node updates from other peers to us
 	in chan *tailcfg.Node
-	// out fans out our node updates to peers
-	out chan *tailcfg.Node
+	// out fans out our node updates to peers connected to us
+	out chan *overlayMessage
 }
 
 func (r *Wasm) IPs() []netip.Addr {
@@ -93,9 +107,11 @@ func (r *Wasm) PickDERPHome(ctx context.Context) error {
 	if report.PreferredDERP == 0 {
 		r.HumanLogf("Failed to determine overlay DERP region, defaulting to %s.", cliui.Code("NYC"))
 		r.DerpRegionID = 1
+		r.DerpLatency = report.RegionLatency[1]
 	} else {
 		r.HumanLogf("Picked DERP region %s as overlay home", cliui.Code(r.DerpMap.Regions[report.PreferredDERP].RegionName))
 		r.DerpRegionID = uint16(report.PreferredDERP)
+		r.DerpLatency = report.RegionLatency[report.PreferredDERP]
 	}
 
 	return nil
@@ -103,6 +119,7 @@ func (r *Wasm) PickDERPHome(ctx context.Context) error {
 
 func (r *Wasm) ClientAuth() *ClientAuth {
 	return &ClientAuth{
+		Web:                  true,
 		OverlayPrivateKey:    r.PeerPriv,
 		ReceiverPublicKey:    r.SelfPriv.Public(),
 		ReceiverDERPRegionID: r.DerpRegionID,
@@ -113,19 +130,32 @@ func (r *Wasm) Recv() <-chan *tailcfg.Node {
 	return r.in
 }
 
-func (r *Wasm) Send() chan<- *tailcfg.Node {
-	return r.out
+func (r *Wasm) SendTailscaleNodeUpdate(node *tailcfg.Node) {
+	r.out <- &overlayMessage{
+		Typ:  messageTypeNodeUpdate,
+		Node: *node.Clone(),
+	}
+}
+
+func (r *Wasm) SendWebrtcCandidate(peer string, cand webrtc.ICECandidateInit) {
+
+	fmt.Println("go: sending webrtc candidate")
+	r.out <- &overlayMessage{
+		Typ:             messageTypeWebRTCCandidate,
+		WebrtcCandidate: &cand,
+	}
 }
 
 type Peer struct {
-	ID   int32
+	ID   string
 	Name string
 	IP   netip.Addr
+	Type string
 }
 
-func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
+func (r *Wasm) Connect(ctx context.Context, ca ClientAuth, offer webrtc.SessionDescription) (Peer, error) {
 	derpPriv := key.NewNode()
-	c := derphttp.NewRegionClient(derpPriv, func(format string, args ...any) {}, netmon.NewStatic(), func() *tailcfg.DERPRegion {
+	c := derphttp.NewRegionClient(derpPriv, logger.Logf(r.HumanLogf), netmon.NewStatic(), func() *tailcfg.DERPRegion {
 		return r.DerpMap.Regions[int(ca.ReceiverDERPRegionID)]
 	})
 
@@ -134,32 +164,36 @@ func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
 		return Peer{}, err
 	}
 
-	sealed := r.newHelloPacket(ca)
+	sealed := r.newHelloPacket(ca, offer)
 	err = c.Send(ca.ReceiverPublicKey, sealed)
 	if err != nil {
 		return Peer{}, fmt.Errorf("send overlay hello over derp: %w", err)
 	}
 
-	updates := make(chan *tailcfg.Node, 2)
+	updates := make(chan *overlayMessage, 8)
 
-	peerID := rand.Int32()
-	r.peers.Store(peerID, updates)
+	old := r.activePeer.Swap(&updates)
+	if old != nil {
+		close(*old)
+	}
 
 	go func() {
-		defer r.peers.Delete(peerID)
+		defer r.activePeer.CompareAndSwap(&updates, nil)
+		defer c.Close()
 
 		for {
 			select {
 			case <-ctx.Done():
 				return
-			case node := <-updates:
-				msg := overlayMessage{
-					Typ:  messageTypeNodeUpdate,
-					Node: *node,
+			case msg, ok := <-updates:
+				if !ok {
+					c.Close()
+					return
 				}
+
 				raw, err := json.Marshal(msg)
 				if err != nil {
-					panic("marshal node: " + err.Error())
+					panic("marshal overlay msg: " + err.Error())
 				}
 
 				sealed := ca.OverlayPrivateKey.SealTo(ca.ReceiverPublicKey, raw)
@@ -175,6 +209,7 @@ func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
 	waitHello := make(chan struct{})
 	closeOnce := sync.Once{}
 	helloResp := overlayMessage{}
+	helloSrc := key.NodePublic{}
 
 	go func() {
 		for {
@@ -191,7 +226,7 @@ func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
 					continue
 				}
 
-				res, _, ovmsg, err := r.handleNextMessage(ca.OverlayPrivateKey, ca.ReceiverPublicKey, msg.Data)
+				res, _, ovmsg, err := r.handleNextMessage(msg.Source, ca.OverlayPrivateKey, ca.ReceiverPublicKey, msg.Data)
 				if err != nil {
 					fmt.Println("Failed to handle overlay message:", err)
 					continue
@@ -207,6 +242,7 @@ func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
 
 				if ovmsg.Typ == messageTypeHelloResponse {
 					helloResp = ovmsg
+					helloSrc = msg.Source
 					closeOnce.Do(func() {
 						close(waitHello)
 					})
@@ -217,17 +253,26 @@ func (r *Wasm) Connect(ctx context.Context, ca ClientAuth) (Peer, error) {
 
 	select {
 	case <-time.After(10 * time.Second):
+		c.Close()
 		return Peer{}, errors.New("timed out waiting for peer to respond")
 	case <-waitHello:
-		updates <- r.lastNode.Load()
+		updates <- &overlayMessage{
+			Typ:  messageTypeNodeUpdate,
+			Node: *r.lastNode.Load(),
+		}
 		if len(helloResp.Node.Addresses) == 0 {
 			return Peer{}, fmt.Errorf("peer has no addresses")
 		}
 		ip := helloResp.Node.Addresses[0].Addr()
+		typ := "cli"
+		if ca.Web {
+			typ = "web"
+		}
 		return Peer{
-			ID:   peerID,
+			ID:   helloSrc.String(),
 			IP:   ip,
 			Name: helloResp.HostInfo.Username,
+			Type: typ,
 		}, nil
 	}
 }
@@ -243,7 +288,7 @@ func (r *Wasm) ListenOverlayDERP(ctx context.Context) error {
 		return err
 	}
 
-	// node priv -> derp priv
+	// node pub -> derp pub
 	peers := xsync.NewMapOf[key.NodePublic, key.NodePublic]()
 
 	go func() {
@@ -252,14 +297,13 @@ func (r *Wasm) ListenOverlayDERP(ctx context.Context) error {
 			select {
 			case <-ctx.Done():
 				return
-			case node := <-r.out:
-				r.lastNode.Store(node)
-				raw, err := json.Marshal(overlayMessage{
-					Typ:  messageTypeNodeUpdate,
-					Node: *node,
-				})
+			case msg := <-r.out:
+				if msg.Typ == messageTypeNodeUpdate {
+					r.lastNode.Store(&msg.Node)
+				}
+				raw, err := json.Marshal(msg)
 				if err != nil {
-					panic("marshal node: " + err.Error())
+					panic("marshal overlay msg: " + err.Error())
 				}
 
 				sealed := r.SelfPriv.SealTo(r.PeerPriv.Public(), raw)
@@ -273,12 +317,14 @@ func (r *Wasm) ListenOverlayDERP(ctx context.Context) error {
 					}
 					return true
 				})
-				// range over peers that we have connected to
-				r.peers.Range(func(key int32, value chan<- *tailcfg.Node) bool {
-					fmt.Println("sending node to outbound peer")
-					value <- node.Clone()
-					return true
-				})
+				if selectedPeer := r.activePeer.Load(); selectedPeer != nil {
+					// *selectedPeer <- &overlayMessage{
+					// 	Typ:  messageTypeNodeUpdate,
+					// 	Node: *msg.Node.Clone(),
+					// }
+					*selectedPeer <- msg
+					fmt.Println("sending message")
+				}
 			}
 		}
 	}()
@@ -291,13 +337,15 @@ func (r *Wasm) ListenOverlayDERP(ctx context.Context) error {
 
 		switch msg := msg.(type) {
 		case derp.ReceivedPacket:
-			res, key, _, err := r.handleNextMessage(r.SelfPriv, r.PeerPriv.Public(), msg.Data)
+			res, key, _, err := r.handleNextMessage(msg.Source, r.SelfPriv, r.PeerPriv.Public(), msg.Data)
 			if err != nil {
 				r.HumanLogf("Failed to handle overlay message: %s", err.Error())
 				continue
 			}
 
-			peers.Store(key, msg.Source)
+			if !key.IsZero() {
+				peers.Store(key, msg.Source)
+			}
 
 			if res != nil {
 				err = c.Send(msg.Source, res)
@@ -310,7 +358,7 @@ func (r *Wasm) ListenOverlayDERP(ctx context.Context) error {
 	}
 }
 
-func (r *Wasm) newHelloPacket(ca ClientAuth) []byte {
+func (r *Wasm) newHelloPacket(ca ClientAuth, offer webrtc.SessionDescription) []byte {
 	var (
 		username string = r.username
 		hostname string = "wush.dev"
@@ -322,7 +370,8 @@ func (r *Wasm) newHelloPacket(ca ClientAuth) []byte {
 			Username: username,
 			Hostname: hostname,
 		},
-		Node: *r.lastNode.Load(),
+		Node:              *r.lastNode.Load(),
+		WebrtcDescription: &offer,
 	})
 	if err != nil {
 		panic("marshal node: " + err.Error())
@@ -332,15 +381,17 @@ func (r *Wasm) newHelloPacket(ca ClientAuth) []byte {
 	return sealed
 }
 
-func (r *Wasm) handleNextMessage(selfPriv key.NodePrivate, peerPub key.NodePublic, msg []byte) (resRaw []byte, nodeKey key.NodePublic, _ overlayMessage, _ error) {
+func (r *Wasm) handleNextMessage(derpPub key.NodePublic, selfPriv key.NodePrivate, peerPub key.NodePublic, msg []byte) (resRaw []byte, nodeKey key.NodePublic, _ overlayMessage, _ error) {
 	cleartext, ok := selfPriv.OpenFrom(peerPub, msg)
 	if !ok {
 		return nil, key.NodePublic{}, overlayMessage{}, errors.New("message failed decryption")
 	}
 
 	var ovMsg overlayMessage
+	fmt.Println(string(cleartext))
 	err := json.Unmarshal(cleartext, &ovMsg)
 	if err != nil {
+		fmt.Printf("Unmarshal error: %#v\n", err)
 		panic("unmarshal node: " + err.Error())
 	}
 
@@ -368,22 +419,62 @@ func (r *Wasm) handleNextMessage(selfPriv key.NodePrivate, peerPub key.NodePubli
 		r.HumanLogf("%s Received connection request from %s", cliui.Timestamp(time.Now()), cliui.Keyword(fmt.Sprintf("%s@%s", username, hostname)))
 		// TODO: impl
 		r.onNewPeer.Invoke(map[string]any{
-			"id":   js.ValueOf(0),
-			"name": js.ValueOf(""),
-			"ip":   js.ValueOf(""),
+			"id":   js.ValueOf(derpPub.String()),
+			"name": js.ValueOf("test"),
+			"ip":   js.ValueOf("1.2.3.4"),
 			"cancel": js.FuncOf(func(this js.Value, args []js.Value) any {
 				return nil
 			}),
 		})
+
+		if ovMsg.WebrtcDescription != nil {
+			r.handleWebrtcOffer(derpPub, &res, *ovMsg.WebrtcDescription)
+		}
+
 	case messageTypeHelloResponse:
 		if !ovMsg.Node.Key.IsZero() {
 			r.in <- &ovMsg.Node
 		}
+
+		if ovMsg.WebrtcDescription != nil {
+			r.onWebrtcAnswer.Invoke(js.ValueOf(derpPub.String()), map[string]any{
+				"type": js.ValueOf(ovMsg.WebrtcDescription.Type.String()),
+				"sdp":  js.ValueOf(ovMsg.WebrtcDescription.SDP),
+			})
+		}
+
 	case messageTypeNodeUpdate:
 		r.HumanLogf("%s Received updated node from %s", cliui.Timestamp(time.Now()), cliui.Code(ovMsg.Node.Key.String()))
 		if !ovMsg.Node.Key.IsZero() {
 			r.in <- &ovMsg.Node
 		}
+
+	case messageTypeWebRTCOffer:
+		res.Typ = messageTypeWebRTCAnswer
+		r.handleWebrtcOffer(derpPub, &res, *ovMsg.WebrtcDescription)
+
+	case messageTypeWebRTCAnswer:
+		r.onWebrtcAnswer.Invoke(js.ValueOf(derpPub.String()), js.ValueOf(map[string]any{
+			"type": js.ValueOf(ovMsg.WebrtcDescription.Type.String()),
+			"sdp":  js.ValueOf(ovMsg.WebrtcDescription.SDP),
+		}))
+
+	case messageTypeWebRTCCandidate:
+		cand := map[string]any{
+			"candidate": js.ValueOf(ovMsg.WebrtcCandidate.Candidate),
+		}
+		if ovMsg.WebrtcCandidate.SDPMLineIndex != nil {
+			cand["sdpMLineIndex"] = js.ValueOf(int(*ovMsg.WebrtcCandidate.SDPMLineIndex))
+		}
+		if ovMsg.WebrtcCandidate.SDPMid != nil {
+			cand["sdpMid"] = js.ValueOf(*ovMsg.WebrtcCandidate.SDPMid)
+		}
+		if ovMsg.WebrtcCandidate.UsernameFragment != nil {
+			cand["usernameFragment"] = js.ValueOf(*ovMsg.WebrtcCandidate.UsernameFragment)
+		}
+
+		r.onWebrtcCandidate.Invoke(derpPub.String(), cand)
+
 	}
 
 	if res.Typ == 0 {
@@ -397,4 +488,36 @@ func (r *Wasm) handleNextMessage(selfPriv key.NodePrivate, peerPub key.NodePubli
 
 	sealed := selfPriv.SealTo(peerPub, raw)
 	return sealed, ovMsg.Node.Key, ovMsg, nil
+}
+
+func (r *Wasm) handleWebrtcOffer(derpPub key.NodePublic, res *overlayMessage, offer webrtc.SessionDescription) {
+	wait := make(chan struct{})
+
+	then := js.FuncOf(func(this js.Value, args []js.Value) interface{} {
+		defer close(wait)
+		desc := args[0]
+
+		fmt.Printf("desc %#v\n", desc)
+		res.WebrtcDescription = &webrtc.SessionDescription{}
+		res.WebrtcDescription.Type = webrtc.NewSDPType(desc.Get("type").String())
+		res.WebrtcDescription.SDP = desc.Get("sdp").String()
+
+		return nil
+	})
+	defer then.Release()
+	catch := js.FuncOf(func(this js.Value, args []js.Value) interface{} {
+		defer close(wait)
+		err := args[0]
+		errStr := err.Call("toString").String()
+
+		fmt.Println("rtc offer callback failed:", errStr)
+		return nil
+	})
+	defer catch.Release()
+
+	r.onWebrtcOffer.Invoke(js.ValueOf(derpPub.String()), map[string]any{
+		"type": js.ValueOf(offer.Type.String()),
+		"sdp":  js.ValueOf(offer.SDP),
+	}).Call("then", then).Call("catch", catch)
+	<-wait
 }
