@@ -1,4 +1,4 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
+import Document, { Html, Head, Main, NextScript } from "next/document";
 
 class MyDocument extends Document {
   render() {
@@ -13,8 +13,9 @@ class MyDocument extends Document {
             rel="stylesheet"
             href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap"
           />
+          <link rel="manifest" href="/app.webmanifest" />
         </Head>
-        <body className="dracula bg-bg text-fg">
+        <body className="bg-bg text-fg">
           <Main />
           <NextScript />
         </body>
